backend/requirements.txt
fastapi
uvicorn
pymongo
requests
python-dotenv
python-multipart
pdfplumber